requirements.txt
customtkinter>=5.2.0
pygame-ce>=2.5.0
sounddevice>=0.4.6
numpy>=1.24.0
openai-whisper
pynput>=1.7.6
pyperclip>=1.8.2
